requirements.txt
boto3==1.34.69
flask==3.0.3
pyyaml==6.0.1
jinja2==3.1.4
botocore==1.34.69
